requirements.txt
jina[standard]==2.1.5
python-pptx==0.6.21
scipy>=1.6.1